fix: next-auth url base path to be dynamic
