+++ containers/ecr-viewer/src/app/signin/components/RedirectButton.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Button } from "@trussworks/react-uswds";
-import { signIn } from "next-auth/react";
+import { SessionProvider, signIn } from "next-auth/react";
 
 import { ProviderDetails } from "@/app/api/auth/auth";
 import { ArrowForward } from "@/app/components/Icon";
@@ -19,7 +19,7 @@ interface RedirectProps {
  */
 export const RedirectButton = ({ provider }: RedirectProps) => {
   return (
-    <>
+    <SessionProvider basePath={`${process.env.BASE_PATH}/api/auth`}>
       <Button
         className="redirect-button"
         type="button"
@@ -30,6 +30,6 @@ export const RedirectButton = ({ provider }: RedirectProps) => {
         Sign in via {provider.name}
         <ArrowForward aria-hidden={true} size={3} />
       </Button>
-    </>
+    </SessionProvider>
   );
 };
